+++ readeck/Domain/UseCase/Settings/SaveSettingsUseCase.swift
@@ -11,6 +11,7 @@ protocol PSaveSettingsUseCase {
     func execute(theme: Theme) async throws
     func execute(urlOpener: UrlOpener) async throws
     func execute(bookmarkSortField: BookmarkSortField, bookmarkSortDirection: BookmarkSortDirection) async throws
+    func execute(disableReaderBackSwipe: Bool) async throws
     func execute(swipeActionConfig: SwipeActionConfig) async throws
 }
 
@@ -98,6 +99,12 @@ final class SaveSettingsUseCase: PSaveSettingsUseCase {
                 bookmarkSortField: bookmarkSortField,
                 bookmarkSortDirection: bookmarkSortDirection
             )
+        )
+    }
+
+    func execute(disableReaderBackSwipe: Bool) async throws {
+        try await settingsRepository.saveSettings(
+            .init(disableReaderBackSwipe: disableReaderBackSwipe)
         )
     }